refactor layers
