+++ lib/data/entities/UserAuth.dart
