+++ lib/domain/repository/user/UserRepositoryRemote.dart
@@ -1,5 +1,5 @@
-import 'package:base_flutter/entities/User.dart';
-import 'package:base_flutter/entities/UserAuth.dart';
+import 'package:base_flutter/data/entities/User.dart';
+import 'package:base_flutter/data/entities/UserAuth.dart';
 
 //Interface
 abstract class UserRepositoryRemote {
