+++ lib/presentation/di/Injector.dart
@@ -1,16 +1,16 @@
-import 'package:base_flutter/local/DatabaseHelper.dart';
-import 'package:base_flutter/local/db/DbUserRepository.dart';
+import 'package:base_flutter/data/local/DatabaseHelper.dart';
+import 'package:base_flutter/data/local/db/UserRepositoryLocalImp.dart';
 import 'package:base_flutter/presentation/views/feed/FeedPresenter.dart';
 import 'package:base_flutter/presentation/views/feedDetail/FeedDetailPresenter.dart';
 import 'package:base_flutter/presentation/views/home/HomePresenter.dart';
 import 'package:base_flutter/presentation/views/login/LoginPresenter.dart';
 import 'package:base_flutter/presentation/views/profile/ProfilePresenter.dart';
 import 'package:base_flutter/presentation/views/users/UserPresenter.dart';
-import 'package:base_flutter/remote/cloud/CloudPostRepository.dart';
-import 'package:base_flutter/remote/cloud/CloudUserRepository.dart';
-import 'package:base_flutter/remote/network/HttpAuth.dart';
-import 'package:base_flutter/usecases/usescase/PostUseCase.dart';
-import 'package:base_flutter/usecases/usescase/UserUseCase.dart';
+import 'package:base_flutter/data/remote/cloud/PostRepositoryRemoteImp.dart';
+import 'package:base_flutter/data/remote/cloud/UserRepositoryRemoteImp.dart';
+import 'package:base_flutter/data/remote/network/HttpAuth.dart';
+import 'package:base_flutter/domain/usescase/PostUseCase.dart';
+import 'package:base_flutter/domain/usescase/UserUseCase.dart';
 import 'package:http/http.dart';
 import 'package:kiwi/kiwi.dart';
 import 'package:sqflite/sqlite_api.dart';
@@ -29,12 +29,12 @@ class Injector {
     mContainer.registerFactory((c)=> database);
     mContainer.registerFactory((c)=> new DatabaseHelper(c.resolve<Database>()));
 
-    mContainer.registerFactory((c) => CloudUserRepository(c.resolve<Client>(), c.resolve<HttpAuth>()));
-    mContainer.registerFactory((c) => CloudPostRepository(c.resolve<Client>()));
-    mContainer.registerFactory((c)=> DbUserRepository(c.resolve<DatabaseHelper>()));
+    mContainer.registerFactory((c) => UserRepositoryRemoteImp(c.resolve<Client>(), c.resolve<HttpAuth>()));
+    mContainer.registerFactory((c) => PostRepositoryRemoteImp(c.resolve<Client>()));
+    mContainer.registerFactory((c)=> UserRepositoryLocalImp(c.resolve<DatabaseHelper>()));
 
-    mContainer.registerSingleton((c) => UserUseCase(c.resolve<CloudUserRepository>(), c.resolve<DbUserRepository>()));
-    mContainer.registerFactory((c) => PostUseCase(c.resolve<CloudPostRepository>()));
+    mContainer.registerSingleton((c) => UserUseCase(c.resolve<UserRepositoryRemoteImp>(), c.resolve<UserRepositoryLocalImp>()));
+    mContainer.registerFactory((c) => PostUseCase(c.resolve<PostRepositoryRemoteImp>()));
 
 
     //Presenter
