+++ lib/presentation/views/components/RowUserProfile.dart
@@ -1,4 +1,4 @@
-import 'package:base_flutter/entities/User.dart';
+import 'package:base_flutter/data/entities/User.dart';
 import 'package:base_flutter/presentation/views/utils/Utils.dart';
 import 'package:base_flutter/presentation/views/components/RoundedImage.dart';
 import 'package:flutter/cupertino.dart';
