+++ lib/presentation/views/feedDetail/FeedDetailView.dart
@@ -1,4 +1,4 @@
-import 'package:base_flutter/entities/Post.dart';
+import 'package:base_flutter/data/entities/Post.dart';
 
 abstract class FeedDetailView {
   void addComment(Comment comment);
