+++ lib/presentation/views/login/LoginPresenter.dart
@@ -1,6 +1,6 @@
 import 'package:base_flutter/presentation/di/Injector.dart';
 import 'package:base_flutter/presentation/views/login/LoginView.dart';
-import 'package:base_flutter/usecases/usescase/UserUseCase.dart';
+import 'package:base_flutter/domain/usescase/UserUseCase.dart';
 import 'package:flutter/foundation.dart';
 
 class LoginPresenter {
