+++ lib/presentation/views/profile/ProfileView.dart
@@ -1,4 +1,4 @@
-import 'package:base_flutter/entities/User.dart';
+import 'package:base_flutter/data/entities/User.dart';
 
 abstract class ProfileView {
   void onCompleteProfile(User user);
